Identify read-depth breakpoints provided a segmentation file

Identify read-depth breakpoints provided a segmentation file

seg.breaks(
  seg,
  fc.pct = 0.2,
  min.seg.size = NULL,
  min.num.probes = NULL,
  low.cov = NULL,
  clean.brk = NULL,
  verbose = TRUE
)

Arguments

seg
(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean

fc.pct
(numeric) copy number change between 2 consecutive segments: i.e (default) cutoff = 0.2 represents a fold change of 0.8 or 1.2

min.seg.size
(numeric) The minimun segment size (in base pairs) to include in the analysis

min.num.probes
(numeric) The minimun number of probes per segment to include in the analysis

low.cov
(data.frame) a data.frame (chr, start, end) indicating low coverage regions to exclude from the analysis

clean.brk
(numeric) identical breakpoints across multiple samples tend to be artifacts; remove breaks > N

Examples

read.depth.breaks()

#> Error in read.depth.breaks(): could not find function "read.depth.breaks"